[jit] _script_compile and _script_class_compile add to the python CU
Actually wire things up so when you do `@torch.jit.script` from Python
things will be put in the Python CU. This removes the hack where we have
to keep the StrongFunctionPtrs alive in Python.

Additional changes that need to be made as part of this:
1. All scripted functions will get their name mangled so that they are
uniquely named in the global CU
2. Classes keep track of their methods
3. Clearing the Python CU involves deleting all classes *and* their
methods.
4. The global python CU was moved into Python, because of static
destruction order issues.

gh-metadata: pytorch pytorch 22724 gh/suo/89/head